removed userauth from shared
as command and query are going to be referencing different data stores

+++ backend/Commands/MoneyTracker.Commands.Application/BudgetService.cs
@@ -1,4 +1,5 @@
-﻿using MoneyTracker.Authentication.Interfaces;
+﻿using MoneyTracker.Authentication.DTOs;
+using MoneyTracker.Authentication.Interfaces;
 using MoneyTracker.Commands.Domain.Entities.BudgetCategory;
 using MoneyTracker.Commands.Domain.Handlers;
 using MoneyTracker.Commands.Domain.Repositories;
@@ -7,19 +8,25 @@ using MoneyTracker.Contracts.Requests.Budget;
 namespace MoneyTracker.Commands.Application;
 public class BudgetService : IBudgetService
 {
-    private readonly IUserAuthenticationService _userAuthService;
     private readonly IBudgetCommandRepository _dbService;
+    private readonly IUserCommandRepository _userRepository;
 
-    public BudgetService(IUserAuthenticationService userAuthService,
-        IBudgetCommandRepository dbService)
+    public BudgetService(
+        IBudgetCommandRepository dbService,
+        IUserCommandRepository userRepository)
     {
-        _userAuthService = userAuthService;
         _dbService = dbService;
+        _userRepository = userRepository;
     }
 
     public async Task AddBudgetCategory(string token, NewBudgetCategoryRequest newBudget)
     {
-        var user = await _userAuthService.DecodeToken(token);
+        var userAuth = await _userRepository.GetUserAuthFromToken(token);
+        if (userAuth == null)
+            throw new InvalidDataException("Token not found");
+        userAuth.ThrowIfInvalid();
+        
+        var user = new AuthenticatedUser(userAuth.User.Id);
         var dtoToDb = new BudgetCategoryEntity(user.Id, newBudget.BudgetGroupId, newBudget.CategoryId, newBudget.Planned);
 
         await _dbService.AddBudgetCategory(dtoToDb);
@@ -27,7 +34,12 @@ public class BudgetService : IBudgetService
 
     public async Task EditBudgetCategory(string token, EditBudgetCategoryRequest editBudgetCategory)
     {
-        var user = await _userAuthService.DecodeToken(token);
+        var userAuth = await _userRepository.GetUserAuthFromToken(token);
+        if (userAuth == null)
+            throw new InvalidDataException("Token not found");
+        userAuth.ThrowIfInvalid();
+        
+        var user = new AuthenticatedUser(userAuth.User.Id);
         var dtoToDb = new EditBudgetCategoryEntity(user.Id, editBudgetCategory.BudgetCategoryId, editBudgetCategory.BudgetGroupId, editBudgetCategory.BudgetCategoryPlanned);
 
         await _dbService.EditBudgetCategory(dtoToDb);
@@ -35,7 +47,12 @@ public class BudgetService : IBudgetService
 
     public async Task DeleteBudgetCategory(string token, DeleteBudgetCategoryRequest deleteBudgetCategory)
     {
-        var user = await _userAuthService.DecodeToken(token);
+        var userAuth = await _userRepository.GetUserAuthFromToken(token);
+        if (userAuth == null)
+            throw new InvalidDataException("Token not found");
+        userAuth.ThrowIfInvalid();
+        
+        var user = new AuthenticatedUser(userAuth.User.Id);
         var dtoToDb = new DeleteBudgetCategoryEntity(user.Id, deleteBudgetCategory.BudgetGroupId, deleteBudgetCategory.BudgetCategoryId);
 
         await _dbService.DeleteBudgetCategory(dtoToDb);
